LayoutTests/imported/w3c/web-platform-tests/css/css-tables/bounding-box-computation-3-expected.txt

Bounding box computation

Checks that all table-internal boxes are positioned correctly

The first table-row-group should have width and height 0:
The first tbody is empty, and should return an height of 0px. The second one should be 100px tall.


PASS Control test: Table width is 120px
PASS Control test: Table height is 120px
FAIL First (empty) table-row-group is 0px wide assert_equals: expected 0 but got 120
FAIL First (empty) table-row-group is 0px tall assert_equals: expected 0 but got 10
FAIL First (empty) table-row-group should be located at 10px left assert_equals: expected 10 but got 0
FAIL First (empty) table-row-group should be located at 10px top assert_equals: expected 10 but got 0
FAIL Second table-row-group is 100px wide assert_equals: expected 100 but got 120
FAIL Second table-row-group is 100px tall assert_equals: expected 100 but got 110
FAIL Second table-row-group should be located at 10px left assert_equals: expected 10 but got 0
PASS Second table-row-group should be located at 10px top
